First draw and first bag.

+++ src/geomentry/Point.java
@@ -1,8 +1,8 @@
 package geomentry;
 
 public class Point {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
     public static final double EPS = 0.0001;
 
     public Point() {
@@ -10,7 +10,7 @@ public class Point {
         this.y = 0;
     }
 
-    public Point(double x, double y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -25,19 +25,19 @@ public class Point {
     }
 
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 }
